+++ source/state_representation/src/space/cartesian/CartesianState.cpp
@@ -69,6 +69,20 @@ CartesianState CartesianState::operator*(double lambda) const {
   return result;
 }
 
+CartesianState& CartesianState::operator/=(double lambda) {
+  if (std::abs(lambda) < std::numeric_limits<double>::min()) {
+    throw std::runtime_error("Division by zero is not allowed");
+  }
+  lambda = 1.0 / lambda;
+  return this->operator*=(lambda);
+}
+
+CartesianState CartesianState::operator/(double lambda) const {
+  CartesianState result(*this);
+  result /= lambda;
+  return result;
+}
+
 CartesianState CartesianState::copy() const {
   CartesianState result(*this);
   return result;